+++ sdk/src/main/java/com/microsoft/did/sdk/BackupAndRestoreService.kt
@@ -2,29 +2,20 @@
 
 package com.microsoft.did.sdk
 
-import com.microsoft.did.sdk.credential.models.VerifiableCredential
-import com.microsoft.did.sdk.crypto.keyStore.EncryptedKeyStore
 import com.microsoft.did.sdk.datasource.file.JweProtectedBackupFactory
 import com.microsoft.did.sdk.datasource.file.MicrosoftBackupSerializer
-import com.microsoft.did.sdk.datasource.file.RawIdentifierUtility
-import com.microsoft.did.sdk.datasource.file.models.JweEncryptedBackupOptions
+import com.microsoft.did.sdk.datasource.file.models.EncryptedBackupData
 import com.microsoft.did.sdk.datasource.file.models.JweProtectedBackup
-import com.microsoft.did.sdk.datasource.file.models.JweProtectedBackupOptions
+import com.microsoft.did.sdk.datasource.file.models.ProtectedBackupData
 import com.microsoft.did.sdk.datasource.file.models.MicrosoftUnprotectedBackup2020
-import com.microsoft.did.sdk.datasource.file.models.MicrosoftUnprotectedBackupOptions
-import com.microsoft.did.sdk.datasource.file.models.PasswordEncryptedBackupOptions
-import com.microsoft.did.sdk.datasource.file.models.PasswordProtectedBackup
-import com.microsoft.did.sdk.datasource.file.models.PasswordProtectedBackupOptions
+import com.microsoft.did.sdk.datasource.file.models.MicrosoftBackup2020Data
+import com.microsoft.did.sdk.datasource.file.models.PasswordEncryptedBackupData
+import com.microsoft.did.sdk.datasource.file.models.PasswordProtectedBackupData
 import com.microsoft.did.sdk.datasource.file.models.UnprotectedBackup
 import com.microsoft.did.sdk.datasource.file.models.UnprotectedBackupOptions
-import com.microsoft.did.sdk.datasource.file.models.VCMetadata
-import com.microsoft.did.sdk.datasource.file.models.WalletMetadata
-import com.microsoft.did.sdk.datasource.repository.IdentifierRepository
-import com.microsoft.did.sdk.util.controlflow.BadPassword
 import com.microsoft.did.sdk.util.controlflow.Result
 import com.microsoft.did.sdk.util.controlflow.UnknownBackupFormat
 import com.microsoft.did.sdk.util.controlflow.UnknownProtectionMethod
-import com.microsoft.did.sdk.util.controlflow.andThen
 import java.io.InputStream
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -34,9 +25,9 @@ class BackupAndRestoreService @Inject constructor(
     private val jweBackupFactory: JweProtectedBackupFactory,
     private val microsoftBackupSerializer: MicrosoftBackupSerializer
 ) {
-    suspend fun createBackup(options: JweProtectedBackupOptions): Result<JweProtectedBackup> {
+    suspend fun createBackup(options: ProtectedBackupData): Result<JweProtectedBackup> {
         return when (options) {
-            is PasswordProtectedBackupOptions -> {
+            is PasswordProtectedBackupData -> {
                 val backup = createUnprotectedBackup(options.unprotectedBackup)
                 if (backup is Result.Success) {
                     jweBackupFactory.createPasswordBackup(backup.payload, options.password);
@@ -52,7 +43,7 @@ class BackupAndRestoreService @Inject constructor(
 
     private suspend fun createUnprotectedBackup(options: UnprotectedBackupOptions): Result<UnprotectedBackup> {
         return when (options) {
-            is MicrosoftUnprotectedBackupOptions -> {
+            is MicrosoftBackup2020Data -> {
                 microsoftBackupSerializer.create(options)
             }
             else -> {
@@ -65,9 +56,9 @@ class BackupAndRestoreService @Inject constructor(
         return jweBackupFactory.parseBackup(backupFile);
     }
 
-    suspend fun restoreBackup(options: JweEncryptedBackupOptions): Result<UnprotectedBackupOptions> {
+    suspend fun restoreBackup(options: EncryptedBackupData): Result<UnprotectedBackupOptions> {
         return when (options) {
-            is PasswordEncryptedBackupOptions -> {
+            is PasswordEncryptedBackupData -> {
                 when (val backupAttempt = options.backup.decrypt(options.password)) {
                     is Result.Success -> importBackup(backupAttempt.payload)
                     is Result.Failure -> backupAttempt
