+++ sdk/src/main/java/com/microsoft/did/sdk/VerifiableCredentialSdk.kt
@@ -6,7 +6,7 @@
 package com.microsoft.did.sdk
 
 import android.content.Context
-import com.microsoft.did.sdk.datasource.file.models.DifWordList
+import com.microsoft.did.sdk.util.DifWordList
 import com.microsoft.did.sdk.di.DaggerSdkComponent
 import com.microsoft.did.sdk.util.log.DefaultLogConsumer
 import com.microsoft.did.sdk.util.log.SdkLog
