+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/MicrosoftBackup2020Data.kt
@@ -4,7 +4,7 @@ package com.microsoft.did.sdk.datasource.file.models
 
 import com.microsoft.did.sdk.credential.models.VerifiableCredential
 
-class MicrosoftUnprotectedBackupOptions (
+class MicrosoftBackup2020Data (
     val walletMetadata: WalletMetadata,
     val verifiableCredentials: List<Pair<VerifiableCredential, VCMetadata>>
 ) : UnprotectedBackupOptions()
