+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/PasswordProtectedBackupData.kt
@@ -2,7 +2,7 @@
 
 package com.microsoft.did.sdk.datasource.file.models
 
-class PasswordProtectedBackupOptions(
+class PasswordProtectedBackupData(
     val password: String,
     override val unprotectedBackup: UnprotectedBackupOptions
-) : JweProtectedBackupOptions()
+) : ProtectedBackupData()
